+++ migration-library/src/main/kotlin/com/quadient/migration/service/deploy/DesignerDeployClient.kt
@@ -23,6 +23,7 @@ import com.quadient.migration.service.deploy.utility.ResourceType
 import com.quadient.migration.service.deploy.utility.ResultTracker
 import com.quadient.migration.service.ResourcePathProvider
 import com.quadient.migration.service.deploy.utility.ConflictDetectorImpl
+import com.quadient.migration.service.deploy.utility.DeployOrderImpl
 import com.quadient.migration.service.deploy.utility.ProgressReporterImpl
 import com.quadient.migration.service.inspirebuilder.InspireDocumentObjectBuilder
 import com.quadient.migration.service.ipsclient.IpsService
@@ -46,6 +47,7 @@ class DesignerDeployClient(
     postProcess: PostProcessImpl,
     conflictDetector: ConflictDetectorImpl,
     progressReporter: ProgressReporterImpl,
+    deployOrder: DeployOrderImpl,
     documentObjectRepository: DocumentObjectRepository,
     imageRepository: ImageRepository,
     attachmentRepository: AttachmentRepository,
@@ -64,6 +66,7 @@ class DesignerDeployClient(
     postProcess,
     conflictDetector,
     progressReporter,
+    deployOrder,
     resourcePathProvider,
     documentObjectRepository,
     imageRepository,