+++ migration-library/src/main/kotlin/com/quadient/migration/service/deploy/InteractiveDeployClient.kt
@@ -35,6 +35,7 @@ import com.quadient.migration.service.deploy.utility.ResultTracker
 import com.quadient.migration.service.getBaseTemplateFullPath
 import com.quadient.migration.service.ResourcePathProvider
 import com.quadient.migration.service.deploy.utility.ConflictDetectorImpl
+import com.quadient.migration.service.deploy.utility.DeployOrderImpl
 import com.quadient.migration.service.deploy.utility.ProgressReporterImpl
 import com.quadient.migration.service.inspirebuilder.InspireDocumentObjectBuilder
 import com.quadient.migration.service.ipsclient.IpsService
@@ -64,6 +65,7 @@ open class InteractiveDeployClient(
     postProcess: PostProcessImpl,
     conflictDetector: ConflictDetectorImpl,
     progressReporter: ProgressReporterImpl,
+    deployOrder: DeployOrderImpl,
     documentObjectRepository: DocumentObjectRepository,
     imageRepository: ImageRepository,
     attachmentRepository: AttachmentRepository,
@@ -82,6 +84,7 @@ open class InteractiveDeployClient(
     postProcess,
     conflictDetector,
     progressReporter,
+    deployOrder,
     resourcePathProvider,
     documentObjectRepository,
     imageRepository,