+++ migration-library/src/test/kotlin/com/quadient/migration/service/deploy/DeployClientTest.kt
@@ -34,6 +34,7 @@ import com.quadient.migration.service.deploy.utility.ProgressReportItem
 import com.quadient.migration.service.deploy.utility.ResourceType
 import com.quadient.migration.service.inspirebuilder.DesignerDocumentObjectBuilder
 import com.quadient.migration.service.DesignerResourcePathProvider
+import com.quadient.migration.service.deploy.utility.DeployOrderImpl
 import com.quadient.migration.service.ipsclient.IpsService
 import com.quadient.migration.shared.IcmFileMetadata
 import com.quadient.migration.shared.IcmMetadata
@@ -84,6 +85,7 @@ class DeployClientTest {
     val resourcePathProvider = DesignerResourcePathProvider(projectConfig)
     val conflictDetector = ConflictDetectorImpl(documentObjectRepository, imageRepository, attachmentRepository, displayRuleRepository, statusTrackingRepository, resourcePathProvider, InspireOutput.Designer)
     val progressReporter = ProgressReporterImpl(documentObjectRepository, imageRepository, attachmentRepository, displayRuleRepository, documentObjectBuilder, statusTrackingRepository, resourcePathProvider, InspireOutput.Designer)
+    val deployOrder = DeployOrderImpl(documentObjectRepository)
 
     private val subject = DesignerDeployClient(
         projectConfig,
@@ -92,6 +94,7 @@ class DeployClientTest {
         postProcess,
         conflictDetector,
         progressReporter,
+        deployOrder,
         documentObjectRepository,
         imageRepository,
         attachmentRepository,